move getBadgeFacets & helpers from IdeaCollection to Badge
